Fix missing module for --version; remove eval shielding error.
This error resulted from moving the old module family Info::Changelog
to Inspect::Changelog. It only occurred on Gitlab because the module
was satisfied from the system path.

The error was further shielded by two pieces of code. First, the
invocation of lintian to determine the version discarded STDERR upon
error. That was addressed in an earlier commit.

The code affected here was also bracketed by an eval, but it is not
clear that it should be. The existence of the changelog file was
already ascertained with -f, so any failure analyzing it, including a
failure to load the required module, should not be ignored as a mere
inability to guess, but rather raise an error. The eval was removed.