Pull most use of the global variable TAGS from Info::Overrides into Processable::Group.
The global variable TAGS is currently required as an argument to
Info::Overrides. This commit pulls most of the use of that variable
from Info::Overrides into the calling code in Processable::Group.

TAGS is now only needed for the real return values That is because,
currently, the tags found in Info::Overrides are the return value
instead of the override data. Ideally the analysis, and the issuance
of tags, would take place in Processable::Group.

One obstacle is that many fields in the specification of override
files are optional. That makes it hard to separate parsing from
analysis.

Another plan for this module is to store the data returned in MLDBM.
The parsing code would then move to collection/override-file, and
Info::Overrides would look simple like the others that use MLDBM.

A database is not presently used because it is unclear that the class
Tags::Override can be serialized. It will be discontinued in the near
future, however, and open the way to MLDBM.

Gbp-Dch: ignore