Localize global singleton $Lintian::Output::GLOBAL into an object that is passed around.
This is a significant step towards separating the handling of
debugging information from the printing of tags.

For debugging, lintian may be able to use some of Perl's standard
features or commonly used modules. The class Lintian::Output will
probably handle only the printing of tags.

The overarching goal behind this reorganization is to allow tags to be
sorted before printing. For all output formats except 'universal' tags
are currently printed as they appear, which depends on the order in
which checks are run.

In the near future, Lintian output will be sortable and sorted. The
output should not depend on the order in which check are run (checks
may run in parallel soon). Likewise, literal tests should not fail
because one check finished before another.

Gbp-Dch: ignore