Ignore potentially helpful output from intltool-extract in checks/debian/po-debconf.pm.
In case of error, this multi-line information is probably helpful to
the user. It was previously provided via N: lines.

Unfortunately, that approach conflicted with an effort to localize the
global singleton $Lintian::Output::GLOBAL. Passing an $OUTPUT object
is not the right solution. The information should instead be attached
to the tag and perhaps be available elsewhere to those looking for it.

Gbp-Dch: ignore