Provide new Moo role for issued Tags in Lintian::Tag.
Classes relying on this role will be instantiated each time a tag is
issued.  That is how they differ from Tag::Info, which is a class of
tag prototypes.

For the time being, the sole user will be a standard tag class that
concatenates arguments with a space in between, as tags do presently.

In the future, there could be other classes that accept higher order
arguments and convert them to strings inside the tag object. One
benefit would be consistent formatting, but that alone may not be
worth the complexity of a separate class.

Gbp-Dch: ignore