Provide a new class for issued tags that concatenate their arguments with spaces in between.
This class instantiates tags that have been found. That is in contrast
to Tag::Info, which is a class of tag prototypes.  Tags represented by
this class concatenate their arguments with a space in between, as all
tags do presently.

At some point, other classes could accept higher order arguments and
convert them to strings inside the tag object. One benefit would be
consistent formatting, but that alone may not be worth the complexity
of having additional classes.

Gbp-Dch: ignore