In Processable::Group, issue new standard tags for unused overrides; make Processable the bearer of tags.
The new class Tag::Standard instantiates tags that have been found.
That is in contrast to Tag::Info, which is a class of tag prototypes.
Tags represented by the new class concatenate their arguments with a
space in between, as all tags do presently.

This change also enables data providers in Info (which have their
origins in Collect) to issue tags. Once the collection scripts are
fully integrated, there is a seamless way to issue tags throughout
Lintian.

In other words, everything is a check!

Gbp-Dch: ignore