In checks, use new attribute accessors from Processable.
Due to the merging of classes, some attributes in Processable are
available under multiple names. This replaces older-style accessors
like 'pkg_name' with their newer equivalent 'name'. The context, as in
$processable->name, already clarifies what is being named.

Please do not use the '->pkg_' prefix anymore when accessing the
attributes. The aliases will be dropped at some point in the future.