Adjust literal tests for new tag output order.
In Lintian's output, tags are now ordered. With the exception of the
'universal' output format, they previously appeared in the order in
which they were emitted, which was the order in which checks ran.
This commit contains no substantive changes.

Adjusts literal tests for the predetermined tag order.

Most tests use the 'universal' format, which was already ordered.

Gbp-Dch: ignore