Adapt skeletons to separate specifications for building and evaluation of test packages.
Adds a new skeleton 'testing' to be used when preparing the test
environment, but not when building.

Removes parts related to testing from those skeletons used to build
packages.

This commit is part of a series to allow adaptive and partial building
of test packages. It will ultimately solve Bug#935907.

Gbp-Dch: ignore