Provide t/bin/split-desc, a tool to split test specifications into some for building and for evaluation of test packages.
This tool can be used to split test specifications into the new format
that provides separate instructions for building and evaluating test
packages.

The program takes a desc file as its sole argument and can be run from
anywhere. It will split the contents of that directory into two. The
tool can be used in a one-off manner, or on many files with:

    $ find t/tags/ -name 'desc' > files
    $ for desc in `cat files` ; do t/bin/split-desc $desc ; done

The splitting cannot happen in one step as part of 'find' (-exec)
because the 'desc' name was retained for the evalation specifications,
and 'find' gets confused about that.

This commit is part of a series to allow adaptive and partial building
of test packages. It will ultimately solve Bug#935907.

Gbp-Dch: ignore