Another attempt to fix the Gitlab build failure; run on equivalent directory in package sources.
The phone goal of fixing the permissions runs before any files in
./root are copied from ROOT_DIR to builddir. The builddir would have
been empty when the previous rules was executed. This should operate
on the files before they are copied with 'cp -rp'.

The previous commit probably did not change anything. The make command
is issued in builddir, so that ./root and $(rootdir) pointed at the
same non-existing location. Now the rule runs on $(ROOT_DIR)/root.

The condition looks like it was reverted, but it wasn't: The negation
made the desired code flow work under the OR operator.

Gbp-Dch: ignore