Add new classification tags for packages maintained by individuals, groups and teams.
The tags are called:

- package-is-team-maintained
- package-is-co-maintained
- package-is-maintained-by-individual

A package is labeled team-maintained if either the Maintainer or the Uploaders
field contains a mailing list.

Packages tagged as co-maintained are maintained by an informal group of people.
This is considered to be the case if there is at least one person listed in the
Uploaders field and the package is not team-maintained. Everything else is
considered to be a package maintained by an individual.

These new tags are inspired by <https://trends.debian.net>.

Signed-off-by: Chris Lamb <lamby@debian.org>