Add XML::LibXML to runtime prerequisites; remove XML::Simple.
While resolving Bug#926543, a decision was made to replace XML::Simple
in Lintian. The problem was caused by a backend in LibXML, but
XML::Simple has additional drawbacks related to the safety of its data
structures and did not offer access to the 'no_network' option.

A ticket was filed under RT#8046 to also add the new runtime
prerequisite on lindsay.d.o.

More information may be available at the Debian bug above or at:

    https://rt.cpan.org/Public/Bug/Display.html?id=131145

XML::Simple was removed as a prerequisite in d/control, but not on
lintian.d.o. That will happen later.