For d/copyright, take files from unpatched orig_index when outside ./debian. (Closes: #844274)
Rewrites the check for d/copyright such that files outside ./debian
are matched from the unpatched orig_index, while files inside ./debian
are taken from the patched index.

As a result, directories that mentioned in d/copyright but emptied by
patches do not raise 'wildcard-matches-nothing-in-dep5-copyright'. In
many cases, it will avoid a 'unused-file-paragraph-in-dep5-copyright',
as well.

The next commit contains a false positive test case that was used to
verify the basic functionality of the change.