Issue new tag gratuitous-manpage when a manpage is not accompanied by an executable. (Closes: #583125)
For now, the tag is issued only for sections 1 (general commands) and
8 (system administration). The other sections look more complicated.
Perhaps section 6 (games) could be added here.

All existing tests for manpages were adjusted except for one. The test
manpages-dependencies ships a manpage in a package called pecho-data.
In that case, the tag was added as expected.

It is unclear how common it is to ship manpages in a separate -data
package. Policy 12.1 clearly discourages it. I struggled with that
question, and then decided to implement the tag.

I was further unsure about the proper severity. The current setting,
minor, is the lowest setting to trigger a warning. That way, most
people (including the reporter for Bug#583125) are still alerted to
extra manpages.