Adjust test for new locations of obsolete DES routines in glibc >= 2.29-5.
This may fix the build failure in the Gitlab pipeline for unstable. A
recent upload of glibc made it even harder to find the symbol and link
to it. That caused the test binaries-obsolete-des to FTBFS.

More information may be available at:

    https://lists.debian.org/debian-lint-maint/2019/12/msg00081.html
    https://lists.debian.org/debian-glibc/2019/12/msg00021.html

Thanks to Aurelien Jarno for suggesting the changes!

According to the control file for glibc 2.29-5, libc6-dev depends on
libcrypt1-dev. No extra build prerequisites were added to the test.