Eliminate a NULL in cachefilesd after failing to parse changelog version.
This probably affects several other packages, as well. I had used the
NULL to indicate that Collect::Source->changeloge_version was unable
to parse the version, but like so often that was a bad idea.

Provides sane and reasonable default values for version components and
also returns a valid, although not necessary meaningful, object of type
Inspect::Changelog::Version when failing to parse a changelog version.

Gbp-Dch: ignore