Add new check for unsafe mailcap entries. (Closes: #33486)
This check parses mailcap files line by line. It emits a tag if the
placeholder (%s) is used with quotes.

As detailed in these references, quoting should be left to the program
that uses the files. Otherwise, it is considered unsafe.

    Bug#33486
    Bug#90483
    Bug#745141
    http://bugs.debian.org/745141#17
    https://lists.debian.org/debian-user/2005/04/msg01185.html
    https://tools.ietf.org/rfc/rfc1524.txt

The bug is from 1999. The mailcap specification is from 1993. The
quoting question was presumably considered with some care.

Only a few programs still ship mailcap files.