Add new tag for when repackaged sources are not properly advertised.
This commit was prompted by Bug#471537. The patch submitted there no
longer worked with the current source tree, but some relevant ideas
were extracted. Lintian now issues a warning for empty orig sources.

The check for repackaged source was transfered to debian/copyright. It
now makes sure the version string reflects a repackaging.

The idea to check for the rules target get-origtargz was not adopted,
as uscan is now widely used for automated repacking. This commit
series does not check the contents of watch files.