Drop tag 'source-contains-empty-directory', a nuisance ignored by many. (Closes: #907727)
This tag was determined to be a nuisance after a careful discussion in
the bug.  Across the archive, it was triggered more than 7,000 times,
with 83 overrides. As the evidence showed, maintainers either could
not do anything about it (if upstream abandoned the project) or were
too embarrassed to ask.

Furthermore, many upstream authors ship empty folders. It is fair to
conclude that the tag existed primarily because some of our tools have
issues with empty directories. The tag description, now deleted,
admitted as much:

    Keeping the empty directory in the source package can prevent
    others from contributing to the package when using tools like
    git-buildpackage(1).

Those problems should be addressed in our toolchain. It was not right
to blame upstream.

Removes the tag, and the code in the cruft check that issued it.