Add new check debian/rules/dh-sequencer; issue a tag when dh sequencer is not used.
This currently issues only the new tag no-dh-sequencer. More
dh-related functionality will be transfered from debian/rules in the
near future.

The present commit does not address all aspects of the new tag
discussed in the bug filing. That is why the bug is not being closed.

The tag name is different and shorter than discussed. Also, the
severity is wishlist (I), even though some advocates had embraced the
higher level of normal (W).

An attempt was made to bypass Haskell packages. They are identified by
looking for the strings "include /usr/share/cdbs/1/class/hlibrary.mk"
or "DEB_CABAL_PACKAGE" in d/rules.

A brand new check file was started to make space for lots of new
features and tags related to the dh sequencer.

This commit is part of a series to introduce more features and tags
surrounding the dh sequencer. It was prompted by Bug#930679 and is a
work in progress.