Allow boolean-false directory components in link targets. (Closes: #892127)
Contrary to hints about link resolution in the initial report, this
was a genuine bug when resolving relative link targets.

An offending package, libaddresses0_0.4.8-3+b1_amd64.deb, and probably
many other packages in GNUstep, used a subdirectory '0' to indicate a
zero version. That directory was resolved incorrectly when part of a
link target.

The relevant routine aborted early because the component evaluated to
boolean false when there were in fact additional components to
process. Specifically,

    usr/lib/libAddresses.so.0

was resolved to

    usr/lib/GNUstep/Frameworks/Addresses.framework/Versions

when the correct result was

    usr/lib/GNUstep/Frameworks/Addresses.framework/Versions/0/libAddresses.so.0

This commit delivers the correct path and resolves the bug.

Thanks to Yavor Doganov from the GNU Project for taking the time to
make the initial report!