Do not consider manpages from related packages when looking for manpages without executables. (Closes: #946471)
The immediate cause of the bug was that section numbers were not
recorded for manpages provided in related packages.  The key to
resolving the bug, however, was the realization that those manpages
should not be considered when looking for ophaned manpages. They are
only considered to when looking for undocumented executables.

To simplify the list comparisons, this commit uses List::Compare to
implement some of the logic in the manpages check.

As a side note, it is not clear that links should be excluded when
requiring that manpages for commands in /usr/sbin are in section 8,
not 1. The check currently looks only at regular files.