Look for COM, not COFF, binaries when warning against Windows executables in sources. (Disengages: #913867)
When COFF binaries are used on Microsoft Windows platforms (mostly
NT), they are called PE32 or similar. The regex still catches those.

For some reason, 274 font metric files from texlive-lang were
misidentified by the 'file' command (see bug below) as COFF, which in
turn caused Lintian to warn against the purported executables. While
the identification remains faulty, COFF files are also not Windows
executables. That fact allowed an effective mitigation.

As a side effect, the change means that Lintian may not warn against
some COFF files in source, but it does not look to be a common issue
from the list of tags on lintian.d.o. Hopefully the magic information
will be adjusted before Lintian is asked to again reject all COFF
executables.

This commit was prompted by Bug#913867, but does not close it. The bug
is assigned to file(1) where it should be solved. This commit merely
mitigates the impact on Lintian.