Bar untested tags from Lintian.
All new tags will have to appear in the test suite. This requirement
was a long-standing design goal and became necessary when splitting up
the complex check 'files'.

The test suite already gained several such defects when the check
'fields' was broken up, but the likelihood is now far greater.