Show obsolete commands in modprobe.d by names; adjust regex.
This improves the check files/modprobe in two ways: First, it
generates a list of unique commands that are obsolete and should not
be used anymore. They are then displayed by name.

Second, it adds an /m modifier to the regex to process input files
line by line, which may be needed for the ^ anchor to work as
intended.

Gbp-Dch: ignore