From test files-general, remove parts that can be tested separately.
The Lintian test suite strives to be more specific in its tests so
that tags can more easily be moved from check to check. To that end,
the test suite simplified the creation of new tests.  Unfortunately,
only a few of the changes are documented.

This reverts those parts of commit 52021d29 that can easily be tested
separately.

Gbp-Dch: ignore