Change test prerequisite to libcrypt-dev, following a recent change in glibc.
The fix to Bug#946396 broke our commit afb4535d. We therefore also
adjust that commit according a similar change in glibc. The hope is
that this commit restores the Gitlab build on unstable.

The relevant glibc commit can be found here:

    https://salsa.debian.org/glibc-team/glibc/commit/b6fa329b0e576d1cb11bdd88df6b7df7bde402e0

Adjusts the prerequisite for the test binaries/binaries-obsolete-des
to libcrypt-dev from libcrypt1-dev.