Look at binaries from entire group when finding unused manpages. (Closes: #949063)
This required the addition of new functionality in Processable::Group.
Manpages, when packaged separately, are prerequisites to the package
with the executables. That makes the analysis asymmetrical.

The check now looks for manpages in all packages that are pulled in
(from the group). For surplus manpages, the check looks at packages
that rely on the present processable (from the group).