Add contents of data/override/renamed-tags to tag declarations.
The previous names are now listed in the field Renamed-From. Multiple
earlier names did not occur here, but can be separated by whitespace.

Some tags did not exist. The rename operation referred to in

    testsuite-unnecessary-testsuite-autopkgtest-header => testsuite-unnecessary-testsuite-autopkgtest-field

seems to have occurred in commit 4960b883, but references the test and
not the tag. (The tags lack the leading 'testsuite-'.) Instead, the
prior alias unnecessary-testsuite-autopkgtest-header was added to the
tag unnecessary-testsuite-autopkgtest-field.

Likewise, the entry

    fields-co-maintained-package-with-no-vcs-headers => fields-co-maintained-package-with-no-vcs-fields

seems to refer to a tag added in commit b5101b37 and renamed in commit
4960b883. Here, co-maintained-package-with-no-vcs-headers was added as
a prior alias to the tag co-maintained-package-with-no-vcs-fields.

The entry

    source-copyright-license-header => source-copyright-license-field

seemed to relate solely to renamed test and was not added anywhere.
The related modifications occurred also in commit 4960b883 but, as far
as the source-copyright check was concerned, involved only changes in
the descriptions and no tag name changes.

Commit 4960b883 introduced other worthwhile tag name changes that
probably should have been memorialized two years ago. No one seems to
have missed these entries. There is no reason to add them now.