In check scripts, remove code that emits classification for debhelper-autoscript-in-maintainer-scripts.
For many tools, this tag includes a version. That leads to frequent
build failures in the test suite when debhelper or another tool was
upgraded.

The solution is to filter the version from the test results. That is
easier when the check stands separate. This removes the code from the
check scripts before it is added elsewhere.

Gbp-Dch: ignore