Add new check maintainer-scripts/generated for the tag debhelper-autoscript-in-maintainer-scripts.
For many tools, this tag includes a version. That leads to frequent
build failures in the test suite when debhelper or another tool was
upgraded.

The solution is to filter the version from the test results. That is
easier when the check stands separate. This creates a new check for
the affected code.

Gbp-Dch: ignore