Associate tag debhelper-autoscript-in-maintainer-scripts with check maintainer-scripts/generated.
For many tools, this tag includes a version. That leads to frequent
build failures in the test suite when debhelper or another tool was
upgraded.

The solution is to filter the version from the test results. That is
easier when the check stands separate. This associates the tag with
the new check.

Gbp-Dch: ignore