In tests, remove unnecessary python2 references. (See: #936952)
In preparation for the release-wide removal of python2, remove some
references to it from the test suite.

The references that are removed here were generated when test cases
were duplicated for each applicable check. The references that are
removed here fulfill no relevant function for the Lintian test suite.

Gbp-Dch: ignore