Free up lab space unconditionally and as part of regular processing. (See: #924449)
The service lintian.d.o continues to suffer from unspecified
out-of-space problems. There is a presumption that they are caused by
lab files remaining behind.

These lab files, which are needed for Lintian's analysis of packages,
should be removed during global destruction unless the command-line
'keep-lab' was specified. This commit attempts to remove any such
conditions. The removal should also occur during regular processing,
to avoid possible issues with irregular program termination.

This commit renders the 'keep-lab' option ineffective.

Also moves the initialization of the 'keep' attribute to a place where
it is not used uninitialized. That change has no effect here. It is
intended to prevent bugs if and when the 'keep-lab' functionality is
reintroduced.

This commit was not tested on lintian.d.o. It may free up space there,
or it may not.