In tests, update expected tags for new tag formats in check debian/watch.
This commit also reflects some procedural changes. For example,
processing is aborted when no version is found.

That explains why the tag debian-watch-does-not-check-gpg-signature in
disappeared in some circumstances. Also, the classification tag
debian-watch-file-standard is no longer issued when a version is
undeclared; previously the extra information shown was empty.

Gbp-Dch: ignore