In non-native template, set watch file version to 3.
The previous version was blank, which triggered a tag for a missing
version. That caused the check to abort early. In at least one test,
the check did not detect a possible upstream key verification because
of it.

This change leads to more useful tags being detected in the test
suite.

Gbp-Dch: ignore