Rewrite check debian/watch. (See: #950250, #950277)
The two bugs that prompted this rewrite are probably caused by the
single line:

    @opts = split($standard >= 4 ? '\s*,\s*' : ',', $opts);

Both bugs could probably be solved by upgrading to the files presented
to watch file version 4, by removing the space between the option
arguments, or by modifying the line above.

To allow the easy addition of new features, this commit rewrite the
check. The central theme is to avoid the line-by-line processing that
made the code hard to understand. The new tag format does not produce
line numbers. It outputs the offending text.

The rewrite seems to have fixed a few features. For example, now the
tag debian-watch-uses-insecure-uri shows more frequently when
appropriate.

Watch files are generally small and only a few lines long. They are
now slurped.

Gbp-Dch: ignore