Remove data/watch-file/version; introduces unnecessary complexity.
The Lintian::Data class is hard to understand and introduces abstract
complexity into simple procedural decisions. In some cases that may be
warranted, but it is not when managing two simple version numbers used
in a single check.

The same functionality is now provided by a single conditional. The
data file is no longer needed, and hereby removed.

The data pattern probably made sense at one point, as it makes it
easier to update some data, especially automatically, but over time
the facilities of Lintian::Data became so complicated that its use
should be discouraged.

Gbp-Dch: ignore