Merge check files/doxygen into files/documentation.
The two checks share one more tag names. After a review, it did not
make sense to rename one use. The tags do the same thing and belong
into the the same check.

Combines the two checks.

This commit is part of an effort to fully partition the old files
check.

Gbp-Dch: ignore