In check copyright, do not look at orig index for native packages.
The collection src-orig-index links to the regular index for native
packages. That will change in the near future. Only the check
copyright relies on it.

In the future, the orig index will only provide data for non-native
packages. With this commit, the copyright check looks at the regular
unpacked index for native packages.

This commit is part of an effort to clean up file name handling in
various indices. It is causing several bugs such as Bug#951136 and
Bug#929729.

Gbp-Dch: ignore