Read files indices from MLDBM databases instead of gzipped flat files.
Index data was previously stored in flat files that were then gzipped.
While that conserved disk space, it caused long-standing and virtually
intractable problems with irregular filenames.

Over the years, filters for newlines and unicode sequences were added
in various places, yet none of the solutions worked perfectly. With
this commit, file names and all attributes, such as ownership, are
stored in simple and fast MLDBM databases.

All file processing is now concentrated in the collections and can be
refined there. The plan is to switch to tar's quoting style 'c' and
then convert file names, and perhaps other data, into proper Perl
strings using the CString mechanism from Data::ParseBinary.

This commit is part of an effort to clean up file name handling in
various indices. It is causing several bugs such as Bug#951136 and
Bug#929729.

Gbp-Dch: ignore