Parse index for orig archives using tar's C quoting style and new regular expressions. (Closes: #950589)
This commit may introduce a new era for file indices at Lintian. For
time immemorial, we parsed plain tar output and compensated for its
shortfalls.

Using Perl's Archive::Tar::File was never an option because it loads
the entire archive into memory. Many large packages, such as the Linux
kernel, should not be handled that way.

With this commit, we use a relatively new (but hopefully not too new)
quoting style offered by recent versions of GNU tar. The current
selection is described here:

    https://www.gnu.org/software/tar/manual/html_node/quoting-styles.html

After some investigation, we settled on the C quoting style. That
decision was driven primarily by the availability of a conversion
facility in Data::ParseBinary. We use the CString converter.

Now only time can tell how this change will fare in Debian and its
derivatives. In the short term, it could create a lot of problems for
people.

For example, other file indices in Lintian have not been converted
(but they will be shortly). I believe the other indices currently do
not unescape file names. For a transition period there could be
mismatches when two indices are compared, such as in debian/copyright.

There are also some indications that the new parser may not handle
symbolic links properly. More details are in a nearby commit.

This commit is part of an effort to clean up file name handling in
various indices. It causes bugs such as Bug#951136 and Bug#929729.

An earlier commit converted all file indices to fast MLDBM databases
instead of flat files. Compared to the gzipped indices, however, the
databases use up more space. One reason is that they use serialized
data formats (currently from Storable). In the long run, something
like this change is most certainly a good idea.

As an added bonus, this commit also uses a new parsing technique for
tar output. Lines are split along the size/date/time axis, which will
never present unexpected whitespaces. It should preserve whitespaces
in the ownership fields, which caused Bug#895175 and Bug#950589, which
is closed here.

As has been noted, the technique is not foolproof, but what is the
likelihood of an ownership looking like this?

    /(\d+)\s+\d{4}-\d{2}-\d{2}\s+\d{2}\:\d{2}(?:\:\d{2})?/

Please bear with Lintian during the period of transition. We strive to
address all bugs in reasonable time frames.