For binary control files, parse index using tar's C quoting style and new regular expressions.
This change is similar to the commit immedidately prior. The parsing
of tar's file listing is improved. It should now be more robust.

As one noteworthy change, tar is no longer invoked with the option
'--preserve-order'. Whatever the original purpose, it is no longer
needed. The results are now stored in a hash (the file path from the
archive serves as the key), and any order is lost, anyway.

This commit is part of an effort to clean up file name handling in
various indices. It is causing several bugs such as Bug#951136 and
Bug#929729.