Allow numerical owner 0/0 when complaining about improper file ownership.
The logic that a numerical owner (or group) of 0 means root was
previously hidden in the handling of file indices. This commit allows
numerical ownership 0/0 as well as root/root when analyzing files. It
also modified the tag message to reflect it.

In tests, two expected tags changed because of it.

An upcoming commit that restores more fidelity to index handling and
requires this change. The exception that 0/0 is functionally
equivalent to root/root for tar files listing only numerical owners is
now handled explicitly.

This commit is part of an effort to clean up file name handling in
various indices. It is causing several bugs such as Bug#951136 and
Bug#929729.

Gbp-Dch: ignore