Convert File::Path to a Moo-based class; respect definition elsewhere.
The object definiton for File::Path was in some places ambigiuous
('children' was a method and a hash key, but with different meanings).
This converts File::Path to a Moo-based class.

Aside from the checks, which use only instance methods, the sole
consumer of the class is File::Index. That class took great liberty in
side-stepping accessors and adding hash keys that were not oficially
part of the File::Path class.

The code in File::Index was changed to respect the File::Path object
definition. Unofficial attributes were added.

This commit is part of an effort to clean up file name handling in
various indices. It is causing several bugs such as Bug#951136 and
Bug#929729.

Gbp-Dch: ignore