Also offer true fidelity for file names in index for patched source trees.
This index is the last to be converted. Multiple NULL terminators
ensure that paths from the output of 'find' correspond to their true
names in the file system.

The strategy employed here differs because a 'find' command is used
for stat-related values. All the other file indices use tar.

Next, we will try to get rid of the temporary databases, They drag
down performance. The plan is to keep the data in memory while the
checks run in parallel (which they currently do not).

Gbp-Dch: ignore