In check, do not sort hard links when printing tag; order is predetermined in File::Index.
Hard links are calibrated in File::Index. In a group, all point to the
entry with the shortest path. Reverse sorting would have been more
appropriate.

Sorting here also makes no sense. It may misrepresent the relationship
between the files, which is indicated by an arrow.

Lintian uses an arrow it treats hard links internally very much like
symbolic links, in the sense that there is a master entry.

This commit is an example for why Lintian's maintainers should enforce
more quality control. Often, checks are written by experts. They come
in a wide variety of styles. A review process might be a good idea.