Move all roles from Lintian::Info into Lintian:Processable.
The data structures previously known as Collect and Info data are now
Moo roles. They can be activated for any kind of Processable. The data
structures have turned out to be useful and are probably here to stay.

This commit moves them closer to the Processable class. The location
in Lintian::Info was a historical holdover.

Gbp-Dch: ignore