For directories, File::Path->basename no longer returns a trailing slash.
This feature was rarely used, and sometimes the slash was stripped.
Brings the behavior in line with that of other basename-like
functions.

A best effort was made to adjust conditionals everywhere for the
possibility that a directory could now pass as regular file. Some
places already restricted to files only.

The edits elsewhere seemed to matter only in the check cruft, where an
algorithm to strip the dreaded slash malfunctioned and required
repair.