In check debian/copyright, use only the 'patched' index, and not the 'orig' index, to match files. (Closes: #951760)
For native packages, the old indices automatically linked the orig
index to the patched index. In some circumstances, the resulting code
was not very clear. For that reason, the indices are no longer linked.

Native packages should only look at the patched index.Most checks only
do that, but some, like debian/copyright and origtar, look at the
upstream tarball as well.

Commit 0a072bc1 tried to adjust the check debian/copyright, but was
incomplete. (That is the only check affected.) Native sources were not
properly analyzed.

For the test package pkg-perl-tools, it meant that 64 of its 132 files
were unaccounted for. It explains the various tags that appeared as
false positives.

After this fix is applied, lintian reports only one tag. Please report
if that is also a false positive.

$ frontend/lintian pkg-perl-tools_0.57.dsc
W: pkg-perl-tools source: file-without-copyright-information .mailmap

Thanks to gregor hermann for reporting this bug!