Eliminate the NULL from ->link attribute in File::Path.
This attribute now returns EMPTY when the object is not a link (and
not NULL). It wasn't the preferred way to test for the presence of a
link, but some people used it that way.

Also some people saw it necessary to change the return value with
constructs like:

    my $link = $file->link // EMPTY;

That is no longer necessary. The ->link attribute will not return
'undef' any more.

Gbp-Dch: ignore