Provide method for all ->descendants in File::Path and use it; retire breadth-first recursion.
The new method is File::Path->descendants. It returns an array of all
descendants of the object, or the empty list if there are none.

The method ->children was modified to return only direct descendants.
There were only two recursion algorithms (and one did not recurse). It
made more sense to split the methods.

Throughout Lintian, all calls to the method ->children that requested
the now retired breadth-first recursion were replaced with calls to
the new method ->descendants.

The packages pkg-perl-tools, pkg-js-tools and xdeb were examined. They
did not use breadth-first recursion and do not need to be modified.