In File::Path and all invocations, rename method ->fs_path with ->unpacked_path.
This commit is part of an effort to provide method names that less
cryptic and, hopefully, self-explanatory.

The old name here was an abbreviation but otherwise not difficult to
figure out. The new name is better, however. It additionally makes
clear that files have to be unpacked in order to be available.

Not all files are unpacked. For example, files from the orig index are
never unpacked (although automatic scanning for licenses may require
it in the future).

Replaces the method call in all parts of Lintian. The packages xdeb,
pkg-perl-tools, and pkg-js-tools did not use the old method, and no
replacement was made.