Rename method ->file_contents with ->slurp in File::Path and all invocations.
The old name was fine, but the standard for this functionality across
Perl seems to be 'slurp.' figure out.

Replaces the method call in all parts of Lintian. The packages xdeb,
pkg-perl-tools, and pkg-js-tools did not use the old method, and no
replacement was made.

Also converted some calls of the format

    path($file->unpacked_path)->slurp;

to

    $file->slurp;

The goal of this change is to encourage slurping in checks. Soon,
checks will run in parallel. The memory used in that check will be
freed when the check ends.

In the near future, there will be fewer reasons to avoid slurping
small files, when otherwise convenient.