In check nodejs, use JSON::MaybeXS instead of plain JSON.
The prerequisite on libjson-perl may prevent libcpanel-json-xs-perl
from being installed automatically even though that version is
preferred and listed as the first alternative.

Instead uses JSON::MaybeXS, which is a drop-in replacement. (The check
does not use the two features, true and false, that differ.)

The prerequsite on libjson-perl will be dropped soon.

A corresponding merge request will be filed with pkg-js-tools in the
near future.