Simplify reference structure in file indices; recursive implementation of resolve_path.
Reimplements path resolution using a recursive algorithm. There is a
hard limit on following links (currently, 18). A bug in this rewrite
could stir up a lot of problems. The job was done with great care, but
requires further testing.

In a finer point, the commit makes the index object available to all
entries. In general, routines now resolve references via the index
instead of following references.

For example, the root entry was previously located via:

    my $current = $self;
    while (my $next = $current->parent_dir) {
        $current = $next;
    }
    return $current;

Now the recursion is no longer necessary:

    $self->index->lookup; (returns base directory on EMPTY arg)

There are opportunities for optimization. For example, child lookups
could be cached or perhaps linked by object, as they were previously,
rather than by name.

The changes presented here have the potential for explosive breakage.
As an upside, the code is easier to understand, and hopefully easier
to debug in the future. Please file bug your reports!