Deal properly with dot from find utility; base directory information for patched source was probably incorrect.
The find utility uses just a single dot to indicate the base
directory. The existing code, which was based on tar output, assumed
it would be ./ when it was not.

Adds an extra filtering step to make sure the directory ends up being
resolved with the empty string in the end.

This bug affected the index for patched sources only. All the others
rely on tar output, which I believe uses ./ for the base directory.

Also added a single-line modifier to accomodate newlines in file
names. Only the first one is really needed, but it is good practice
throughout Lintian from now on.

Gbp-Dch: ignore